# api/openapi.yaml
openapi: 3.0.3
info:
  title: GophKeeper API
  version: 1.0.0
  description: API stores client-encrypted opaque records.
servers:
  - url: https://localhost:8080
paths:
  /health:
    get:
      summary: Check service health
      responses:
        "200":
          description: Service is healthy
  /v1/register:
    post:
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/login:
    post:
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "401":
          $ref: "#/components/responses/Error"
  /v1/items:
    get:
      summary: List encrypted records and deletion tombstones
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Error"
  /v1/items/{id}:
    put:
      summary: Create, replace, or delete an encrypted record
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutItem"
      responses:
        "200":
          description: Stored record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Error"
        "409":
          description: Version conflict; body is the current record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Credentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 8
    Token:
      type: object
      required: [token]
      properties:
        token:
          type: string
    PutItem:
      type: object
      required: [version]
      properties:
        ciphertext:
          type: string
        version:
          type: integer
          format: int64
          minimum: 0
        deleted:
          type: boolean
    Item:
      allOf:
        - $ref: "#/components/schemas/PutItem"
        - type: object
          required: [id, updated_at]
          properties:
            id:
              type: string
            updated_at:
              type: string
              format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string